--- src/main/java/nextstep/support/NotExistEntityException.java
@@ -1,4 +1,0 @@
-package nextstep.support;
-
-public class NotExistEntityException extends RuntimeException {
-}
