+++ src/main/java/nextstep/support/exception/NotExistEntityException.java
@@ -1,0 +1,10 @@
+package nextstep.support.exception;
+
+public class NotExistEntityException extends RuntimeException {
+    public NotExistEntityException() {
+    }
+
+    public NotExistEntityException(String message) {
+        super(message);
+    }
+}
